cli_auto_up: Partly fix handling of mountpoints
The cli_auto_up command right now does not properly handle
mountpoints.

Fix the case when going up one level inside of a mounted
module.

Note that this does not yet fix the case where the current level
is the top-level of the mounted module. In that case the up command
should switch to the presence container that holds the mounted
module. Instead it switches all the way up to the topmost level, so
it has the same effect as the top command.