xml_merge1(): Allow merging trees with nodes below mountpoints
xml_merge1() fails right now when merging trees which have data
below mountpoints with an error message like this:

XML node %s/%s has no corresponding yang specification (Invalid XML or wrong Yang spec?

This is because xml_merge1() does not handle mountpoints. Adjust
it accordingly.

Note that this was encountered when generating the state of a a yang
module that has both state and config below a mountpoint. In that case
netconf_trymerge() will be called to merge state and config and will
trigger the problem in xml_merge1().